Added ability to choose which renderer you would like to use for each field.
It is not remembered across page changes at the moment.

I had to change the test data module so that it doesn't really extend the standard table page/application anymore. This was because of class loading issues. The rmi class loader couldn't find the StringWrapper class because it was starting to look for it in the table.api module as that is where the DataFacility was located. I will have to some how tell it to use a different class loader.